refactor(game-engine): split choice-handlers — extract handleRerollChoose (S27.8.14)
`handleRerollChoose` (~199 lines including the private helpers
`getLonerThreshold` and `consumeTeamReroll`) is the largest handler
in `choice-handlers.ts`. Extract it into the new module
`actions/reroll-choose-handler.ts` to drop `choice-handlers.ts` below
the secondary DoD target of <= 400 lines.

Re-export from `choice-handlers.ts` to preserve the API consumed by
`actions.ts` and the test suites.

No cycle back to `actions.ts` nor `choice-handlers.ts`: the handler
only calls `mechanics/*`, `core/*`, `utils/*`, `skills/*`,
`actions/failure-helpers`, `actions/ball-pickup`.

Orphan import cleanup in `choice-handlers.ts` (4 removed): `Player`,
`isInOpponentEndzone`, `awardTouchdown`, `samePos`.

Additional cleanup: `_targetNumber` (legacy destructuring artifact
never used) removed from the new module's `pendingReroll`
destructuring.

`choice-handlers.ts` shrinks 517 -> 288 lines (-229 net lines,
**below the DoD secondary threshold of 400 lines**).

5 new smoke tests in `reroll-choose-handler.test.ts` cover the export
from the new module, the re-export from `choice-handlers.ts`, the
`pendingReroll` absent gates, and state immutability.

Coverage exclusion added in `vitest.config.ts` (same v8 source-map
quirk as all other S27.8.x extractions).

- 4928/4928 game-engine tests pass (+5 new smoke tests)
- 1787/1787 server tests pass (no regression)
- typecheck + lint + build clean
- coverage 25.33% functions/branches (>= 25% threshold)
