+++ packages/game-engine/src/actions/choice-handlers.ts
@@ -2,7 +2,7 @@
  * S27.8.7 — Handlers de choix (suite a un blocage / poussee /
  * follow-up / relance d'equipe) extraits de `actions.ts`.
  *
- * Quatre handlers cohesifs autour des phases de choix utilisateur
+ * Trois handlers cohesifs autour des phases de choix utilisateur
  * apres un evenement initial :
  *  - `handleBlockChoose` : choix du resultat de des de bloc apres
  *    un blocage qui a roule plusieurs des (l'attaquant ou le
@@ -11,21 +11,24 @@
  *    celles disponibles (Fend, Frenzy, chain push).
  *  - `handleFollowUpChoose` : confirmation du suivi (follow-up)
  *    apres une poussee classique.
- *  - `handleRerollChoose` : decision sur l'utilisation d'une
- *    relance d'equipe pour un jet rate (dodge / GFI / pickup),
- *    avec test Solitaire (Loner) le cas echeant.
+ *
+ * S27.8.14 — `handleRerollChoose` (le plus gros, ~199 lignes incluant
+ * `getLonerThreshold` et `consumeTeamReroll`) extrait dans
+ * `actions/reroll-choose-handler.ts` pour ramener ce module sous la
+ * cible secondaire DoD `<= 400 lignes`. Re-exporte ici pour preserver
+ * l'API consommee par `actions.ts` et les tests.
  *
  * Aucune dependance interne au dispatcher : tout passe par les
- * modules deja extraits (`failure-helpers`, `ball-pickup`) +
- * `core/game-state` + `mechanics/*`.
+ * modules deja extraits (`failure-helpers`, `ball-pickup`,
+ * `reroll-choose-handler`) + `core/game-state` + `mechanics/*`.
  */
 
+// S27.8.14 — `Player` consomme uniquement par `getLonerThreshold` qui
+// a migre vers `reroll-choose-handler.ts`. Plus d'import direct ici.
 import type {
   GameState,
-  Player,
   Position,
   RNG,
-  TeamId,
   BlockResult,
 } from '../core/types';
 import {
@@ -40,45 +43,18 @@ import {
 } from '../mechanics/blocking';
 import { isFendActiveForFollowUp } from '../mechanics/fend';
 import { hasFrenzy } from '../mechanics/frenzy';
-import { getArmBarBonus } from '../mechanics/arm-bar';
-import {
-  isInOpponentEndzone,
-  awardTouchdown,
-  checkTouchdowns,
-} from '../mechanics/ball';
-import { samePos } from '../mechanics/movement';
-import { hasSkill } from '../skills/skill-effects';
-import {
-  performDodgeRollWithNotification,
-  performPickupRollWithNotification,
-} from '../utils/dice-notifications';
+// S27.8.14 — `isInOpponentEndzone`, `awardTouchdown`, `samePos`
+// consommes uniquement par `reroll-choose-handler.ts`. Plus d'import
+// direct ici. `checkTouchdowns` reste utilise par `handleBlockChoose`
+// et `handleFollowUpChoose`.
+import { checkTouchdowns } from '../mechanics/ball';
 import { createLogEntry } from '../utils/logging';
-import { applyRollFailure, applyPickupFailure } from './failure-helpers';
-import { handleBallPickup } from './ball-pickup';
 
-/**
- * Retourne le seuil Loner du joueur (3, 4 ou 5) ou `null` s'il n'a
- * pas Loner.
- */
-function getLonerThreshold(player: Player): number | null {
-  if (hasSkill(player, 'loner-3')) return 3;
-  if (hasSkill(player, 'loner-4')) return 4;
-  if (hasSkill(player, 'loner-5')) return 5;
-  return null;
-}
-
-/**
- * Consomme une relance d'equipe et marque le flag `rerollUsedThisTurn`.
- */
-function consumeTeamReroll(state: GameState, team: TeamId): GameState {
-  const newRerolls = { ...state.teamRerolls };
-  if (team === 'A') {
-    newRerolls.teamA = Math.max(0, (newRerolls.teamA ?? 0) - 1);
-  } else {
-    newRerolls.teamB = Math.max(0, (newRerolls.teamB ?? 0) - 1);
-  }
-  return { ...state, teamRerolls: newRerolls, rerollUsedThisTurn: true };
-}
+// S27.8.14 — `handleRerollChoose` (incluant les helpers prives
+// `getLonerThreshold` et `consumeTeamReroll`) extrait dans
+// `actions/reroll-choose-handler.ts`. Re-export pour preserver
+// l'API consommee par `actions.ts`.
+export { handleRerollChoose } from './reroll-choose-handler';
 
 /**
  * Gere le choix de resultat de blocage parmi les des roules.
@@ -309,209 +285,4 @@ export function handleFollowUpChoose(
   }
 
   return checkTouchdowns(newState);
-}
-
-/**
- * Gere le choix de relance d'equipe pour un jet rate (dodge / GFI /
- * pickup), avec test Solitaire (Loner) le cas echeant.
- */
-export function handleRerollChoose(
-  state: GameState,
-  move: { type: 'REROLL_CHOOSE'; useReroll: boolean },
-  rng: RNG,
-): GameState {
-  if (!state.pendingReroll) return state;
-
-  const {
-    rollType,
-    playerId,
-    team,
-    targetNumber: _targetNumber,
-    modifiers,
-    playerIndex,
-    from,
-    to,
-  } = state.pendingReroll;
-  let newState: GameState = { ...state, pendingReroll: undefined };
-
-  if (!move.useReroll) {
-    // Relance refusee : appliquer les consequences de l'echec
-    const declineLog = createLogEntry(
-      'action',
-      `Relance refusée`,
-      playerId,
-      team,
-    );
-    newState.gameLog = [...newState.gameLog, declineLog];
-
-    if (rollType === 'pickup') {
-      return applyPickupFailure(newState, playerIndex, rng);
-    } else {
-      // dodge ou gfi : chute + turnover
-      return applyRollFailure(newState, playerIndex, rng);
-    }
-  }
-
-  // Verification Loner : le joueur doit reussir un jet D6 >= seuil
-  const lonerPlayer = newState.players[playerIndex];
-  const lonerThreshold = getLonerThreshold(lonerPlayer);
-  if (lonerThreshold !== null) {
-    const lonerRoll = Math.floor(rng() * 6) + 1;
-    const lonerSuccess = lonerRoll >= lonerThreshold;
-    const lonerLog = createLogEntry(
-      'dice',
-      `Solitaire (${lonerThreshold}+) : ${lonerRoll}/${lonerThreshold} ${
-        lonerSuccess ? '✓' : '✗'
-      }`,
-      playerId,
-      team,
-      {
-        diceRoll: lonerRoll,
-        targetNumber: lonerThreshold,
-        success: lonerSuccess,
-      },
-    );
-    newState.gameLog = [...newState.gameLog, lonerLog];
-
-    if (!lonerSuccess) {
-      // Loner check failed: team reroll is consumed (wasted), original
-      // failure applies
-      newState = consumeTeamReroll(newState, team);
-      const wastedLog = createLogEntry(
-        'action',
-        `Relance d'équipe gaspillée (Solitaire raté)`,
-        playerId,
-        team,
-      );
-      newState.gameLog = [...newState.gameLog, wastedLog];
-
-      if (rollType === 'pickup') {
-        return applyPickupFailure(newState, playerIndex, rng);
-      } else {
-        return applyRollFailure(newState, playerIndex, rng);
-      }
-    }
-  }
-
-  // Relance acceptee : consommer la relance
-  newState = consumeTeamReroll(newState, team);
-  const rerollLog = createLogEntry(
-    'action',
-    `Relance d'équipe utilisée !`,
-    playerId,
-    team,
-  );
-  newState.gameLog = [...newState.gameLog, rerollLog];
-
-  if (rollType === 'dodge') {
-    // Relancer le jet d'esquive
-    const dodgeResult = performDodgeRollWithNotification(
-      newState.players[playerIndex],
-      rng,
-      modifiers,
-    );
-    newState.lastDiceResult = dodgeResult;
-    const logEntry = createLogEntry(
-      'dice',
-      `Relance esquive: ${dodgeResult.diceRoll}/${dodgeResult.targetNumber} ${
-        dodgeResult.success ? '✓' : '✗'
-      }`,
-      playerId,
-      team,
-      {
-        diceRoll: dodgeResult.diceRoll,
-        targetNumber: dodgeResult.targetNumber,
-        success: dodgeResult.success,
-      },
-    );
-    newState.gameLog = [...newState.gameLog, logEntry];
-
-    if (dodgeResult.success) {
-      // Touchdown check
-      const mover = newState.players[playerIndex];
-      if (mover.hasBall && isInOpponentEndzone(newState, mover)) {
-        return awardTouchdown(newState, mover.team, mover);
-      }
-      return newState;
-    } else {
-      // Echec apres team reroll : appliquer l'echec avec bonus Arm Bar.
-      const dodger = newState.players[playerIndex];
-      const armBarBonus = from ? getArmBarBonus(newState, dodger, from) : 0;
-      return applyRollFailure(newState, playerIndex, rng, armBarBonus);
-    }
-  } else if (rollType === 'gfi') {
-    // Relancer le jet de GFI
-    const gfiRoll = Math.floor(rng() * 6) + 1;
-    const gfiSuccess = gfiRoll >= 2;
-    const logEntry = createLogEntry(
-      'dice',
-      `Relance GFI: ${gfiRoll}/2 ${gfiSuccess ? '✓' : '✗'}`,
-      playerId,
-      team,
-      { diceRoll: gfiRoll, targetNumber: 2, success: gfiSuccess },
-    );
-    newState.gameLog = [...newState.gameLog, logEntry];
-
-    if (gfiSuccess) {
-      // Touchdown check
-      const mover = newState.players[playerIndex];
-      if (mover.hasBall && isInOpponentEndzone(newState, mover)) {
-        return awardTouchdown(newState, mover.team, mover);
-      }
-      // Ball pickup check
-      if (newState.ball && to && samePos(newState.ball, to)) {
-        return handleBallPickup(newState, mover, rng, playerIndex);
-      }
-      return newState;
-    } else {
-      return applyRollFailure(newState, playerIndex, rng);
-    }
-  } else if (rollType === 'pickup') {
-    // Relancer le jet de pickup
-    const pickupResult = performPickupRollWithNotification(
-      newState.players[playerIndex],
-      rng,
-      modifiers,
-    );
-    newState.lastDiceResult = pickupResult;
-    const logEntry = createLogEntry(
-      'dice',
-      `Relance pickup: ${pickupResult.diceRoll}/${pickupResult.targetNumber} ${
-        pickupResult.success ? '✓' : '✗'
-      }`,
-      playerId,
-      team,
-      {
-        diceRoll: pickupResult.diceRoll,
-        targetNumber: pickupResult.targetNumber,
-        success: pickupResult.success,
-      },
-    );
-    newState.gameLog = [...newState.gameLog, logEntry];
-
-    if (pickupResult.success) {
-      newState.ball = undefined;
-      newState.players[playerIndex] = {
-        ...newState.players[playerIndex],
-        hasBall: true,
-      };
-      const successLog = createLogEntry(
-        'action',
-        `Ballon ramassé avec succès (relance)`,
-        playerId,
-        team,
-      );
-      newState.gameLog = [...newState.gameLog, successLog];
-      // Touchdown check
-      const picker = newState.players[playerIndex];
-      if (isInOpponentEndzone(newState, picker)) {
-        return awardTouchdown(newState, picker.team, picker);
-      }
-      return newState;
-    } else {
-      return applyPickupFailure(newState, playerIndex, rng);
-    }
-  }
-
-  return newState;
 }
